Inverted Index commit 1.6-开发完成

+++ custom-inverted-index/src/main/java/com/swj/es/study/index/SearchStats.java
@@ -10,5 +10,5 @@ import lombok.Data;
 @Data
 public class SearchStats {
     private int documentsSearched;
-    private int durationNanos;
+    private long durationNanos;
 }
